# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gym Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List fitness classes
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: trainerId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FitnessClass"
    post:
      operationId: createClass
      tags: [classes]
      summary: Create a fitness class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassBody"
      responses:
        "201":
          description: Class created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FitnessClass"
  /classes/{id}:
    get:
      operationId: getClass
      tags: [classes]
      summary: Get class by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FitnessClass"
        "404":
          description: Not found
    patch:
      operationId: updateClass
      tags: [classes]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClassBody"
      responses:
        "200":
          description: Class updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FitnessClass"
    delete:
      operationId: deleteClass
      tags: [classes]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── BOOKINGS ───────────────────────────────────────────────────────────────
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List class bookings
      parameters:
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking (cancel, confirm, etc.)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Booking updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Cancel booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cancelled

  # ─── PERSONAL TRAINING SESSIONS ─────────────────────────────────────────────
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: memberId
          in: query
          schema:
            type: ["integer", "null"]
        - name: classId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AttendanceRecord"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAttendanceBody"
      responses:
        "201":
          description: Attendance recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceRecord"

  # ─── ANALYTICS ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    FitnessClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
        trainerName:
          type: ["string", "null"]
        capacity:
          type: integer
        enrolledCount:
          type: integer
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        location:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - capacity
        - enrolledCount
        - scheduledAt
        - durationMinutes
        - category
        - status
        - createdAt
        - updatedAt
    CreateClassBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
        capacity:
          type: integer
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        location:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string
      required:
        - name
        - capacity
        - scheduledAt
        - durationMinutes
        - category
        - status
    UpdateClassBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        trainerId:
          type: ["integer", "null"]
        capacity:
          type: integer
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        location:
          type: ["string", "null"]
        category:
          type: string
        status:
          type: string

    # ─── BOOKING ────────────────────────────────────────────────────────────
    Booking:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: ["string", "null"]
        classId:
          type: integer
        className:
          type: ["string", "null"]
        status:
          type: string
        bookedAt:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - memberId
        - classId
        - status
        - bookedAt
        - createdAt
        - updatedAt
    CreateBookingBody:
      type: object
      properties:
        memberId:
          type: integer
        classId:
          type: integer
      required:
        - memberId
        - classId
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string

    # ─── PT SESSION ─────────────────────────────────────────────────────────
...
    AttendanceRecord:
      type: object
      properties:
        id:
          type: integer
        memberId:
          type: integer
        memberName:
          type: ["string", "null"]
        classId:
          type: ["integer", "null"]
        className:
          type: ["string", "null"]
        checkedInAt:
          type: string
        checkedOutAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - memberId
        - checkedInAt
        - createdAt
    CreateAttendanceBody:
      type: object
      properties:
        memberId:
          type: integer
        classId:
          type: ["integer", "null"]
        checkedInAt:
          type: string
        checkedOutAt:
          type: ["string", "null"]
      required:
        - memberId
        - checkedInAt

    # ─── ANALYTICS ──────────────────────────────────────────────────────────
...