# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Booking {
  id: number;
  memberId: number;
  /** @nullable */
  memberName?: string | null;
  classId: number;
  /** @nullable */
  className?: string | null;
  status: string;
  bookedAt: string;
  createdAt: string;
  updatedAt: string;
}