# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBookingBody {
  memberId: number;
  classId: number;
}