# lib/api-zod/src/generated/types/fitnessClass.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Gym Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface FitnessClass {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  trainerId?: number | null;
  /** @nullable */
  trainerName?: string | null;
  capacity: number;
  enrolledCount: number;
  scheduledAt: string;
  durationMinutes: number;
  /** @nullable */
  location?: string | null;
  category: string;
  status: string;
  createdAt: string;
  updatedAt: string;
}